+++ srcs/output/output_char_str.c
@@ -6,7 +6,7 @@
 /*   By: jmarsal <jmarsal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/02 01:06:21 by jmarsal           #+#    #+#             */
-/*   Updated: 2016/11/04 16:29:31 by jmarsal          ###   ########.fr       */
+/*   Updated: 2016/11/10 12:29:04 by jmarsal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ static void	print_str(t_result *result, size_t i)
 {
 	if (I_L_CONV == 's')
 	{
-		if (I_STRLEN >= PRECISION_O)
+		if ((int)I_STRLEN >= PRECISION_O)
 			ft_buffer_add(RET_STR, RET_STR->len, I_STR, ft_strlen(I_STR) - PRECISION_O);
 		else
 			ft_buffer_add(RET_STR, RET_STR->len, I_STR, ft_strlen(I_STR));
