+++ srcs/output/output_helper.c
@@ -6,7 +6,7 @@
 /*   By: jmarsal <jmarsal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/27 22:53:18 by jmarsal           #+#    #+#             */
-/*   Updated: 2016/11/07 15:50:19 by jmarsal          ###   ########.fr       */
+/*   Updated: 2016/11/10 15:52:49 by jmarsal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,23 +15,7 @@
 void	get_specifier_and_ajust_width(t_result *result, char specifier)
 {
 	L_CONV = specifier;
-	// if (L_CONV != 'c' && L_CONV != 's')
-	// {
-	// 	if (A_PLUS == 1)
-	// 		I_WIDTH_CPY--;
-	// 	if (A_SPACE == 1)
-	// 		I_WIDTH_CPY--;
-	// }
 	STRLEN = ft_strlen(STR);
-	// I_WIDTH_CPY -= STRLEN;
-	// I_PRECISION_CPY -= STRLEN;
-	// if (((I_IS_WIDTH && I_IS_PRECISION && A_MINUS && A_PLUS) ||
-	// 	(I_IS_WIDTH && I_IS_PRECISION && A_MINUS && A_SHARP)) &&
-	// 	I_PRECISION_CPY > 0)
-	// 	I_WIDTH_CPY = (L_CONV == 'x' || L_CONV == 'X') ?
-	// 		I_WIDTH_CPY - I_PRECISION_CPY - 2 : I_WIDTH_CPY - I_PRECISION_CPY;
-	// else if (I_IS_WIDTH && I_IS_PRECISION && A_MINUS && !A_SHARP)
-	// 	I_WIDTH_CPY -= I_PRECISION_CPY;
 }
 
 void	add_padding(t_result *result, size_t i, int size, char c)
@@ -48,8 +32,8 @@ void	add_padding(t_result *result, size_t i, int size, char c)
 	}
 	else if (PRECISION_CPY > 0)
 	{
-		padding = ft_strnew(PRECISION_O - (I_STRLEN));
-		ft_memset(padding, c, PRECISION_O - (I_STRLEN));
+		padding = ft_strnew(PRECISION_O - ((int)I_STRLEN));
+		ft_memset(padding, c, PRECISION_O - ((int)I_STRLEN));
 		ft_buffer_add(RET_STR, RET_STR->len, padding, ft_strlen(padding));
 	}
 	ft_free_null (padding);
